requirements.txt
Flask
tensorflow
numpy
Pillow
scikit-image
matplotlib
keras
split-folders
scikit-learn
seaborn
streamlit